refactor: filter 파일 Global 위치 변경 (#95)
